+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 ### Fixed
 - #130: Do not automatically try to re-login if kicked out from another location
+- The user now gets a helpful error message instead of a popup when trying to send a message beyond server size limits
 
 ## 0.6.0
 ### Added